requirements.txt
dash == 2.3.1
pandas == 1.4.2
dash-bootstrap-components == 1.1.0
plotly == 5.7.0
dash-labs == 1.0.3
psycopg2-binary == 2.9.3
dash_daq == 0.5.0
numpy==1.22.3
geopy==2.2.0
gunicorn==20.1.0
python-dateutil==2.8.2
selenium==4.2.0
webdriver-manager==3.7.0
pandas==1.4.2